HeThongDiemDanh/requirements.txt
Pillow
numpy
thinker
opencv-pyhton
tkcalendar
mysql-connector-python